LPS-158126 Move module declarations to ts_modules
