+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/index.d.ts
@@ -12,4 +12,44 @@
  * details.
  */
 
+import {FieldType} from './utils/dataConverter';
+
 export {FieldFeedback} from './core/components/FieldFeedback';
+export {EVENT_TYPES as FORM_EVENT_TYPES} from './custom/form/eventTypes';
+export {getDDMFormFieldSettingsContext} from './utils/dataConverter';
+
+export function convertToFormData(body: unknown): unknown;
+
+export function makeFetch({
+	body,
+	headers,
+	method,
+	url,
+	...otherProps
+}: {
+	body: unknown;
+	headers?:
+		| {
+				Accept: string;
+		  }
+		| undefined;
+	method?: string | undefined;
+	url: unknown;
+	[x: string]: unknown;
+}): unknown;
+
+export function useConfig(): {
+	fieldTypes: FieldType[];
+	formReportDataURL: string;
+	portletNamespace: string;
+};
+
+export function useForm(): ({
+	payload,
+	type,
+}: {
+	payload?: unknown;
+	type: string;
+}) => void;
+
+export function useFormState<T extends {[key: string]: unknown}>(): T;
