+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/utils/dataConverter.ts
@@ -157,3 +157,76 @@ export function _fromDDMFormToDataDefinitionPropertyName(propertyName: string) {
 			return propertyName;
 	}
 }
+
+interface DataDefinition {
+	customProperties: DataDefinitionCustomProperties;
+	defaultValue?: unknown;
+	fieldType?: unknown;
+	indexType?: unknown;
+	indexable?: unknown;
+	label?: unknown;
+	localizable?: unknown;
+	name?: unknown;
+	nestedDataDefinitionFields: DataDefinition[];
+	readOnly?: unknown;
+	repeatable?: unknown;
+	required?: unknown;
+	showLabel?: unknown;
+	tip?: unknown;
+}
+
+interface DataDefinitionCustomProperties {
+	options?: LocalizedValue<unknown>;
+	[key: string]: unknown;
+}
+
+interface Field<T = unknown> {
+	fieldName: string;
+	localizable?: boolean;
+	localizedValue?: LocalizedValue<T>;
+	multiple?: unknown;
+	nestedFields?: Field[];
+	options?: unknown;
+	settingsContext: {pages: unknown[]};
+	type: FieldTypeName;
+	value: T;
+}
+
+export interface FieldType {
+	label: string;
+	name: FieldTypeName;
+	settingsContext: {pages: unknown[]};
+}
+
+export type FieldTypeName =
+	| 'checkbox_multiple'
+	| 'captcha'
+	| 'checkbox'
+	| 'color'
+	| 'date'
+	| 'document_library'
+	| 'fieldset'
+	| 'grid'
+	| 'help_text'
+	| 'image'
+	| 'key_value'
+	| 'localizable_text'
+	| 'multi_language_option_select'
+	| 'numeric'
+	| 'numeric_input_mask'
+	| 'object_field'
+	| 'object-relationship'
+	| 'options'
+	| 'password'
+	| 'paragraph'
+	| 'redirect_button'
+	| 'radio'
+	| 'rich_text'
+	| 'search_location'
+	| 'separator'
+	| 'select'
+	| 'text'
+	| 'validation';
+
+type Locale = Liferay.Language.Locale;
+type LocalizedValue<T> = Liferay.Language.LocalizedValue<T>;
