+++ modules/apps/data-engine/data-engine-js-components-web/types/src/main/resources/META-INF/resources/js/utils/dataConverter.d.ts
@@ -44,3 +44,75 @@ export declare function getDDMFormFieldSettingsContext({
 export declare function _fromDDMFormToDataDefinitionPropertyName(
 	propertyName: string
 ): string;
+interface DataDefinition {
+	customProperties: DataDefinitionCustomProperties;
+	defaultValue?: unknown;
+	fieldType?: unknown;
+	indexType?: unknown;
+	indexable?: unknown;
+	label?: unknown;
+	localizable?: unknown;
+	name?: unknown;
+	nestedDataDefinitionFields: DataDefinition[];
+	readOnly?: unknown;
+	repeatable?: unknown;
+	required?: unknown;
+	showLabel?: unknown;
+	tip?: unknown;
+}
+interface DataDefinitionCustomProperties {
+	options?: LocalizedValue<unknown>;
+	[key: string]: unknown;
+}
+interface Field<T = unknown> {
+	fieldName: string;
+	localizable?: boolean;
+	localizedValue?: LocalizedValue<T>;
+	multiple?: unknown;
+	nestedFields?: Field[];
+	options?: unknown;
+	settingsContext: {
+		pages: unknown[];
+	};
+	type: FieldTypeName;
+	value: T;
+}
+export interface FieldType {
+	label: string;
+	name: FieldTypeName;
+	settingsContext: {
+		pages: unknown[];
+	};
+}
+export declare type FieldTypeName =
+	| 'checkbox_multiple'
+	| 'captcha'
+	| 'checkbox'
+	| 'color'
+	| 'date'
+	| 'document_library'
+	| 'fieldset'
+	| 'grid'
+	| 'help_text'
+	| 'image'
+	| 'key_value'
+	| 'localizable_text'
+	| 'multi_language_option_select'
+	| 'numeric'
+	| 'numeric_input_mask'
+	| 'object_field'
+	| 'object-relationship'
+	| 'options'
+	| 'password'
+	| 'paragraph'
+	| 'redirect_button'
+	| 'radio'
+	| 'rich_text'
+	| 'search_location'
+	| 'separator'
+	| 'select'
+	| 'text'
+	| 'validation';
+declare type Locale = Liferay.Language.Locale;
+declare type LocalizedValue<T> = Liferay.Language.LocalizedValue<T>;
+export {};
