+++ modules/apps/object/object-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/ts_modules/data-engine-js-components-web.d.ts
@@ -12,4 +12,6 @@
  * details.
  */
 
-declare module 'data-engine-js-components-web';
+declare module 'data-engine-js-components-web' {
+	export * from 'data-engine-js-components-web/src/main/resources/META-INF/resources/js/index';
+}
