+++ modules/apps/object/object-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/ts_modules/dynamic-data-mapping-form-field-type.d.ts
@@ -13,11 +13,5 @@
  */
 
 declare module 'dynamic-data-mapping-form-field-type' {
-	type FieldChangeEventHandler<T = unknown> = (event: {
-		target: {
-			value: T;
-		};
-	}) => void;
-
-	function ReactFieldBase(props: any): JSX.Element;
+	export * from 'dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/index';
 }
